cloud_services/requirements.txt
kafka-python
influxdb
paho-mqtt
